+++ test.js
@@ -19,6 +19,11 @@ equal(
 )
 
 equal(
+  md().use(anchor, { level: [2, 4] }).render('# H1\n\n## H2\n\n### H3\n\n#### H4\n\n##### H5'),
+  '<h1>H1</h1>\n<h2 id="h2">H2</h2>\n<h3>H3</h3>\n<h4 id="h4">H4</h4>\n<h5>H5</h5>\n'
+)
+
+equal(
   md().use(anchor, { permalink: true }).render('# H1'),
   '<h1 id="h1">H1 <a class="header-anchor" href="#h1" aria-hidden="true">¶</a></h1>\n'
 )